week6/pyproject.toml
[tool.poetry]
name = "week6"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
kafka-python = "^2.0.2"
confluent-kafka = "^2.0.2"
requests = "^2.28.2"
avro = "^1.11.1"
fastavro = "^1.7.3"
faust-streaming = "^0.10.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
